openstack create_node takes ex_network_config
there has been a regression in the openstack driver in libcloud where
previously it was possible to use create_node to spin up instances using
pre-existing ports. this broke because:

- **kwargs are no longer accepted
- create_node now assumes that network is Network objects by uuid

if the user wishes to provide an alternative network configuration not
relying on uuids when spinning up a new node, with this commit there is
now an option to pass in a plain dict as ex_network_config instead.

this will allow booting nodes with networks.port, networks.fixed_ip or
networks.tag rather than only the network.uuid assumption that currently
is made by the networks kwarg.